+++ src/index.ts
@@ -4,6 +4,7 @@ import "./headless-navigation"
 import "./headless-disclosure"
 import "./headless-rating"
 import "./headless-darkmode"
+import "./headless-toggle"
 
 import "./headless-button"
 import "./headless-scrolltop"